Don't enforce black for now.
We still have a lot of repos that are going to be running on 3.5, it
would be really annoying for them if they needed a second venv just to
run black.  Especially because it might not always be possible in
various dev/test environments.  We'll just hold off on enforcing this
till next year when we're on at least 3.6