Fix requirements installation
Running `pip install -r requirements/dev.txt` was failing because `-e .`
was expanded to an absolute path. This is a long-standing issue in
pip-tools: https://github.com/jazzband/pip-tools/issues/204

So we got rid of the "-e ." line in dev.in and recompiled the
requirement files. We realized that this did not pull all requirements
because some of them were hardcoded in setup.cfg. We decided to move
those requirements to a base.in file and to dynamically load this file
from setup.py. This removes the need to manually upgrade a few
dependencies. To achieve that, we had to add a MANIFEST.in file.

Then, in order to fix tests we had to include test requirements, so we
created a test.in requirement file.

Note that the `py35-pylint24` still fails because black==20.8b1 is
incompatible with Python 3.5. AFAIK there is no version compatible with
Python 3.5.