Report generic code annotation errors
Previously, it was required to run `code_annotations --lint` in order to verify
code annotations. The errors then had to be collected manually, for instance
with a paver task. This made it intractable to generalize code annotation
checking to multiple IDAs. With this new checker, code annotation errors are
sent to pylint, which makes it easy for everyone to quickly detect annotation
errors. To achieve this, we make use of the new code annotations API created in
v1.1.0.