fix: unit tests with pylint>=2.6.0
The upgrade to pylint 2.6.0 broken the unit tests. Turns out that the
edx_lint.pylint module was no longer being used, so pylint was not reporting
any lint warning.

Also, for some reason, the `ignore-long-lines` option present in pylintrc is
ignored in the unit tests, so we need to pass it explicitely to the test
runner.