build: Remove tox constraint.
The `tox` package is already 6 minor revisions ahead at 4.6.4.  If there
are still plugins that don't support 4.x.x, then they are likely
stagnant and need to be removed or updated ourselves.

However, as long as we keep this constraint here, we can't easily find
and fix those issues.  In many of the cases, this constraint was added
due to the incompatibility of tox-battery with tox 4.x.x.  However,
tox-battery has updated its `install_requires` to be explicit of this
dependency.

https://github.com/signalpillar/tox-battery/blob/master/setup.py#L20

Another issue we're running into is that some of the dependencies of tox
are starting to publish security vulnerabilities.  It's lower risk since
this is in dev and CI but leaving this as is will increase security
noise making it harder to respnod to real signals.

Specifically, tox<4.0.0 depends on a version of `py` which has a
security vulnerability.  Dependabot is picking this up and making some
noise in a lot of our repos.