feat!: Don't allow writing of commitlint.config.js files.
That file is not meant to be overrdiden on a repo by repo basis so it
doesn't make sense that we should be able to write a new one.

Also with the update to the commitlint github action @6 the file needs
to be different suffixed and that is handled in the github action
upstream correctly leaving even fewer reasons to allow this override.

BREAKING CHANGE: edx-lint can no longer write commitlint.config.js
files.  If you have such a local file, it should be removed so that the
correct upstream version of the file can be picked up.