fix: Don't pin django-simple-history
We removed this pin in edx-platform a while ago and if we can run these
migrations there, I think any other services should be able to handle
this as well.  Also there is no ticket with details here and so it is
not clear if there was a specific concern other than performance issues
with adding indexes.

Reference (default branch in edx-platform as of today):
https://github.com/openedx/edx-platform/blob/d29171c04649f29a26bfc96aa73b64bf01bfdd5b/Makefile#L124-L125